backend/services/nlp_agent/pyproject.toml
[project]
name = "nlp-agent"
version = "0.1.0"
description = "Aura Journal NLP Agent Service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-cloud-language>=2.17.2",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "uvicorn[standard]>=0.35.0",
    "pytest>=8.4.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "pymongo>=4.13.2",
    "sqlalchemy>=2.0.41",
    "psycopg2-binary>=2.9.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
